Correct user check for life counter MANAGER-END endpoint

+++ BoardGameGeekLike/Models/Dtos/Response/UsersCheckForLifeCounterManagerEndResponse.cs
@@ -3,6 +3,6 @@
     public class UsersCheckForLifeCounterManagerEndResponse
     {
         public bool IsFinished { get; set; } = false;
-        public int? Duration_minutes { get; set; }  
+        public double? Duration_minutes { get; set; }  
     }
 }
