+++ BoardGameGeekLike/Models/Entities/LifeCounterManager.cs
@@ -29,7 +29,7 @@ namespace BoardGameGeekLike.Models.Entities
 
         public long? EndingTime { get; set; }
 
-        public int? Duration_minutes { get; set; }
+        public double? Duration_minutes { get; set; }
 
         public bool IsFinished { get; set; } = false;
